:bug: fix: parcing to file

+++ src/service/convertFile.service.js
@@ -1,9 +1,11 @@
 import { fileExists, readFileJson, readFileTxt, readFileXml, removeFile } from '../utils/fileHandler.util.js';
 import { convertCsvToListMap, convertJsonToListMap, convertXmlToListMap } from '../utils/mappingData.util.js';
+import { unflattenObject } from '../utils/objectFlatter.util.js';
 import { convertDataToCvs, convertDataToJson, convertDataToTXT, convertDataToXML } from '../utils/dataParser.util.js';
 import FILE_TYPES from '../utils/constants/fileTypes.constant.js';
 import ServiceError from '../utils/errors/service.error.util.js';
 import Upload from '../utils/errors/codes/upload.codes.js';
+import { parse } from 'dotenv';
 
 const extractDataFromFiles = async (filePath, secretKey) => {
     if (!fileExists(filePath)) {
@@ -28,7 +30,8 @@ const extractDataFromFiles = async (filePath, secretKey) => {
             data = readFileTxt(filePath);
             return convertCsvToListMap(data, ',', secretKey);
         default: {
-            removeFile(filePath);
+            if (fileExists(filePath))
+                removeFile(filePath);
             throw new ServiceError(
                 `Unsupported file type: ${fileExtension}`,
                 Upload.UPLOAD_FILE_TYPE_NOT_SUPPORTED
@@ -40,15 +43,16 @@ const extractDataFromFiles = async (filePath, secretKey) => {
 };
 
 const dataConverterToFile = async (data, fileType, delimiter) => {
+    const dataConverted = unflattenObject(data);
     switch (fileType.toLowerCase()) {
         case FILE_TYPES.JSON:
-            return convertDataToJson(data);
+            return convertDataToJson(dataConverted);
         case FILE_TYPES.XML:
-            return convertDataToXML(data);
+            return convertDataToXML(dataConverted);
         case FILE_TYPES.CSV:
-            return convertDataToCvs(data, delimiter);
+            return convertDataToCvs(dataConverted, delimiter);
         case FILE_TYPES.TXT:
-            return convertDataToTXT(data, delimiter);
+            return convertDataToTXT(dataConverted, delimiter);
         default:
             throw new ServiceError(
                 `Unsupported file type: ${fileType}`,
@@ -57,18 +61,30 @@ const dataConverterToFile = async (data, fileType, delimiter) => {
     }
 }
 
+const isConvertionNeeded = (filePath, requestedConvertion) => {
+    const fileExtension = filePath.split('.').pop().toLowerCase();
+    if (fileExtension.toLowerCase() === requestedConvertion.toLowerCase()) {
+        return false;
+    }
+
+    return true;
+}
+
 const parseToNewFile = (data) => {
     return Buffer.from(data, 'utf-8').toString('base64');
 }
 
 const fileParserService = async (filePath, secretKey, fileType, delimiter) => {
-    console.log(`Processing file: ${filePath}`);
     try {
+        if (!isConvertionNeeded(filePath, fileType)) {
+            console.log(`No conversion needed for file: ${filePath}`);
+            return parseToNewFile(readFileTxt(filePath));
+        }
+
         const result = await extractDataFromFiles(filePath, delimiter);
         const parsedData = await dataConverterToFile(result, fileType, delimiter);
-        console.log(`File processed successfully: ${filePath}`);
-        console.log('Encrypted targets in file:', result);
         console.log('Parsed data:', parsedData);
+        console.log('Flattened object:', result);
         return parseToNewFile(parsedData);
     } catch (e) {
         console.error(`Error processing file: ${filePath}`, e);
@@ -77,7 +93,8 @@ const fileParserService = async (filePath, secretKey, fileType, delimiter) => {
             e.code || Upload.PROCESSINGFILE_ERROR,
         )
     } finally {
-        removeFile(filePath);
+        if (fileExists(filePath))
+            removeFile(filePath);
         console.log(`Temporary file removed: ${filePath}`);
     }
 }
